Bring back the local patch that has a habit of vanishing every time we update mesa.
Specifically, we disallow in radeondrm for dri clients mapping
registers, so don't try and map them (and thus fail as we currently
were). for r300+ this was only used for falling back on old drm versions
(doesn't matter). For r100, the new BO abstraction used the SWI number
(in hardware scratch reg 3) for the buffer age, so use the newly added
getparam member to grab that info instead of trying to read the mapped
registers.

Update to the lastest kernel headers  before you even think about
building this or trying to use a snapshot on r100/r200.

So now radeon works with mesa again, hoorah!

Tested on rv250 by Josh Elsasser, and on R420 (and x800) by myself.