disable additional rules to regen sources when python is not found
After a cvs import timestamps change which triggers rules that attempt
to call python to regenerate files.  When python is not found this
generates empty files that override those from the distfile, breaking
the build.

When building inside xenocara python is never found as the pkg paths
are not searched.