Make sure the fontconfig cache in the xbaseXX.tgz set is valid.
Tarballs only store timestamps with 1s resolution, while the
fontconfig cache uses nano-second resolution from the filesystem to
check that a directory hasn't changed.

So the font directories extracted by the installer were almost always
considered out of date, causing applications that use fontconfig to
trigger a rebuild of the cache.

Now, when installing with DESTDIR set, we use a perl one-liner to
remove the nanoseconds from the mtime of the fonts directories before
running fc-cache, so that the timestamps recorded in the cache match
what will be set after extracting the xfont set.

Help and suggestions from at least guenther@, millert@, espie@
and deraadt@