Build Mesa intel and radeon vulkan drivers and amd64 and i386
The intel driver can be used with inteldrm on ivy bridge or newer.
The radeon driver only works with amdgpu not radeondrm.

As we can't use python in xenocara add phony targets to create the
same output as python scripts which create json files for the loader.

This is just the ICDs, to use vulkan the loader and additional ports
are required.

ok mpi@ on an earlier version.  Changed to call shm_unlink()
directly after shm_mkstemp() as suggested by kettenis@