Fix the size of indicators in multi-monitors setups.
ssh-askpass(1) is computing the size of its indicators relatively to
the screen resolution. When multiple monitors are present, the support
for Xinerama computes the dimensions of the window to be created
relatively to the screen on which it will appear. But the computation
of the indicator size is based on the size of the whole display,
resulting in too small indicators (and too many of them if the screens
hare layed out horizontally).

Fix that by computing the resolution of the whole display before
taking xinerama into account.

ok djm@