Retire xman(1).
It's totally broken (requires MLINKS which we removed long ago,
relies on the old man.conf(5) format we stopped supporting long ago,
Alfred Morgan reported additional bugs on bugs@, ...),
it reimplements man(1) in a very poor way,
and it lacks many useful features that man(1) in an xterm(1)
readily provides, so fixing it would be a waste of time.

matthieu@ deraadt@ krw@ agree with the direction.