Remove osrelease from system.fvwmrc
This is the version string that shows up as part of "OpenBSD 6.8" in the
bottom right window in a default installation.

But as this version stems from the build machine's kernel with which fvwm
was build, it can differ from the version you're running on your machine
which is what you'd expect as user.

Such values must not be derived from the build environment;  instead of
fixing it up, remove the version number entirely and leave it to the user's
~/.fvwmrc to update the label if need be.

OK deraadt