util: enable futex usage on BSDs after 7dc2f4788288
From Jan Beich
46c368907fcf333a19881d28c46e997845d00faf in mainline Mesa

We had a related local patch in 19.2 which wasn't needed when 20.1 was
imported as the above commit was backported to the upstream 20.1 branch.
When Mesa 20.0 was imported after issues with 20.1 on Haswell the
changes to use futexes for simple_mtx.h and u_queue.h were lost.

Noticed by otto@ and kettenis@ when looking for memory leaks.