Ignore WSMOUSEIO_GTYPE ioctl failure when checking /dev/wsmouse
When xf86-input-ws has attached separately to all existing mouse
devices already or there are no mice plugged in, xf86-input-ws fails
to attach to the wsmouse mux as a default fallback because there is
no wsmouse device left in the mux to handle the WSMOUSEIO_GTYPE
ioctl.  If a USB mouse is later plugged in, there will not be any
xf86-input-ws driver loaded listening to /dev/wsmouse for input.

ok matthieu