Rewrite the privsep imsg code using the new ibuf API.
Properly wrap the imsg_read() / imsg_get() calls in a while loop.
Do not initialize the fd of the struct ibuf to -1 (this is already done).
Rewrite the imsg parser using imsg_get_ibuf(), ibuf_get and a local new
function ibuf_get_string() which abstracts the extraction of the strings
from the ibuf.
Properly clear and free the user, pass and style variables.

OK tb@ matthieu@